[core] TermCtl: Replace color and mode placeholders in print/format using a custom parser
Do not delegate to fmt. There is a regression in fmt 11, which broke that - named placeholders cannot be mixed with automatically indexed ones.
Although I think this should be allowed, using the named args for the TermCtl color placeholders was basically a hack. A custom parser will be more flexible and allow more freedom regarding the syntax.

The new placeholders are processed purely in runtime. Unknown placeholders are left untouched. Compile-time checking is still possible, but that doesn't seem as important feature. Actual replacing by escape sequences must be done in runtime, as it depends on what is attached to the output stream (tty or pipe).